Literature Review
In this phase of the thesis, research had to be made regarding the foundation of knowledge required to start building up the thesis.  Therefore, an estimate of around 20 references were found on the student popular websites of google scholar and ieeexplore.  These papers ranged from different gamification experiences and methods with different gamification elements and structures to Virtual Reality Simulations built to aid students/users into learning a certain skill/subject.  After reading most of the knowledge found within these papers, the literature review was built out of different comparisons found within each paper discovered.  Different comparisons where made depending on the different prototype methods and testing methods used in their papers.  In addition, all this research would aid me in building a final prototype based on the various attempts other people had in the past.  These were then listed in the different sections and subsections of the literature review chapter for easy referencing when building the prototype and the rest of the thesis.  
This research mostly consisted of Gamification Elements and their effects on Game-Based Learning and Virtual Reality.  Elements explored where the aspect of competition within learning games, feedback and its importance, the rewarding system and how it affects motivation and Virtual Interaction and freedom of experimentation within Learning Simulations.