Methodology
The methodology chapter requires the researcher to figure out and explain, the method they will be using to get the results they need and answer the research questions given out in the SOI.  Firstly, the reader would need a bit of background about the packages and the software used in the creation of the prototype.  To satisfy that, a small paragraph was created explaining.  In order to explain the testing procedures, a decision regarding quantitative and qualitative testing methods would find itself explained within this section of the thesis.  Once that was dealt with the methodology continues to explain various gamification elements and variables that was instantiated in order to generate and receive the results required.  Finally, the methods used to create the prototype were explained and compared to other referenced papers found in the literature review.  In other words, the methodology is explained as a guide for other researchers to follow if they would want to achieve the same results.