Results and Discussion
This chapter keeps track of any gathered results and answers received from the prototype.  The results and discussion in its name will be the chapter where the previous comparisons found in the methodology and literature review will now be compared with the results generated from the prototype in order to take note of any difference in the experiment.  These comparisons will be taken from the google forms survey created for qualitative reasons and the game database built for quantitative requirements.  In this section, tables and possible graphs will be set up to make sure the reader understands the results obtained and whether they show a positive impact or a negative one.  A good constructive discussion was also discussed regarding the results, firearm safety in general and the limitations and observations that one had during the testing procedure.  All this is done while still comparing the results with different referenced papers.  This chapter will also house the answers underlying the research questions within the SOI and will determine the outcome of the main research question as a result of the thesis.