Conclusion, Introduction and Abstract
Conclusion: 
The final chapter would contain a summary of the discussion of the results and make sure that the main research question of the thesis was understood.  This would also be the time to list the strengths and weaknesses of the thesis.  The important aspects that effected the thesis’s result base.  These would include for example the COVID-19 outbreak as a weakness and the prototype’s interaction features as a strength.  With that said, future recommendations would also be found in the conclusion.  
Introduction: 
The introduction starts by identifying the main problem and gives a brief explanation regarding an event or experience.  The motivation regarding the problem should also be written and listed but the most important parts are the research questions and objectives.  These would be the questions that would be answered in the results and discussion.  Finally, a small summary describing the chapters found in the thesis would give the reader a heads up of what to expect from the thesis.  
Abstract: 
The abstract can be simply explained as the dissertation’s summary.  This will contain a paragraph of a mere 500 words of information from each of the chapters highlighting the most interesting parts.